# backend/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: FitnessLab API
  description: API for FitnessLab management system
  version: 1.0.0
servers: []
...
paths:
...
  /api/products:
    post:
      tags:
      - Products
      operationId: createProduct
      summary: Create a new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Product created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductCreationResponse'
    get:
      tags:
      - Products
      operationId: getAllProducts
      summary: Get all products
      responses:
        '200':
          description: List of all products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductView'
  /api/products/downloadFromWix:
    post:
      tags:
        - Products
      operationId: downloadFromWix
      summary: Apply all pending Wix updates to local products
      responses:
        '200':
          description: All Wix updates applied successfully
        '500':
          description: Error applying Wix updates
  /api/products/checkForWixUpdatesForAll:
    post:
      tags:
        - Products
      operationId: checkForWixUpdatesForAll
      summary: Check for updates from Wix for all linked products without applying them
      responses:
        '200':
          description: Check completed successfully
        '500':
          description: Error checking for Wix updates
  /api/products/{productId}:
    get:
      tags:
      - Products
      operationId: getProduct
      summary: Get product by ID
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductView'
        '404':
          description: Product not found
    put:
      tags:
      - Products
      operationId: updateProduct
      summary: Update product information
      parameters:
      - name: productId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProductRequest'
      responses:
        '200':
          description: Product updated successfully
  /api/products/{productId}/uploadToWix:
    post:
      tags:
        - Products
      operationId: uploadToWix
      summary: upload a product to wix
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '201':
          description: Product successfully uploaded to wix
  /api/products/{productId}/downloadFromWix:
    post:
      tags:
        - Products
      operationId: downloadFromWixForProduct
      summary: Download and apply changes from Wix for a specific product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product successfully downloaded from Wix
        '500':
          description: Error downloading product from Wix
  /api/products/{productId}/checkForWixUpdates:
    post:
      tags:
        - Products
      operationId: checkForWixUpdates
      summary: Check if there are updates on Wix for a specific product without applying them
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Check completed successfully
        '404':
          description: Product not found
        '500':
          description: Error checking for Wix updates
...
components:
  schemas:
...
    CreateProductRequest:
      type: object
      required:
      - slug
      - name
      - productType
      - audience
      - requiresMembership
      - pricingVariant
      - behavior
      properties:
        slug:
          type: string
        name:
          type: string
        productType:
          type: string
        audience: &id001
          $ref: '#/components/schemas/Audience'
        requiresMembership:
          type: boolean
        pricingVariant:
          $ref: '#/components/schemas/PricingVariantConfig'
        behavior:
          $ref: '#/components/schemas/ProductBehaviorConfig'
        description:
          type: string
          nullable: true
        termsAndConditions:
          type: string
          nullable: true
        visibility:
          type: string
          enum:
          - PUBLIC
          - HIDDEN
          - ARCHIVED
          default: PUBLIC
        buyable:
          type: boolean
          default: true
        buyerCanCancel:
          type: boolean
          default: true
        perks:
          type: array
          nullable: true
          items:
            type: string
    UpdateProductRequest:
      type: object
      required:
        - slug
        - name
        - productType
        - audience
        - requiresMembership
        - pricingVariant
        - behavior
      properties:
        slug:
          type: string
        name:
          type: string
        productType:
          type: string
        audience:
          $ref: '#/components/schemas/Audience'
        requiresMembership:
          type: boolean
        pricingVariant:
          $ref: '#/components/schemas/PricingVariantConfig'
        behavior:
          $ref: '#/components/schemas/ProductBehaviorConfig'
        description:
          type: string
          nullable: true
        termsAndConditions:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [PUBLIC, HIDDEN, ARCHIVED]
          nullable: true
        buyable:
          type: boolean
          nullable: true
        buyerCanCancel:
          type: boolean
          nullable: true
        perks:
          type: array
          items:
            type: string
          nullable: true
        linkedPlatforms:
          type: array
          items:
            $ref: '#/components/schemas/LinkedPlatformSync'
          nullable: true
    Audience:
      type: string
      enum:
      - INTERNAL
      - EXTERNAL
      - BOTH
    ProductBehaviorConfig:
      type: object
      description: Behavior and business rules applied to Contracts.
      required:
      - canBePaused
      properties:
        canBePaused:
          type: boolean
          description: Whether the contract for this product can be paused.
        renewalLeadTimeDays:
          type: integer
          nullable: true
          description: Number of days before expiry when auto-renewal should be triggered
            in your own logic. Null if not time-based or not applicable.
        maxActivePerCustomer:
          type: integer
          nullable: true
          description: Maximum number of concurrently active contracts of this product
            (or group) allowed per customer. Null means no explicit limit.
        maxPurchasesPerBuyer:
          type: integer
          nullable: true
          minimum: 0
          maximum: 10
          description: Mirrors Wix Plan.maxPurchasesPerBuyer semantics. 0 = unlimited
            purchases, 1 = a customer can only purchase once, up to 10.
        numberOfSessions:
          type: integer
          nullable: true
          description: Total number of sessions associated with this product, if applicable.
    ProductCreationResponse:
      type: object
      properties:
        productId:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
    ProductView:
      type: object
      properties:
        productId:
          type: string
        slug:
          type: string
        name:
          type: string
        productType:
          type: string
        audience:
          type: string
        requiresMembership:
          type: boolean
        pricingVariant:
          $ref: '#/components/schemas/PricingVariantConfig'
        behavior:
          $ref: '#/components/schemas/ProductBehaviorConfig'
        description:
          type: string
          nullable: true
        termsAndConditions:
          type: string
          nullable: true
        visibility:
          type: string
          enum:
          - PUBLIC
          - HIDDEN
          - ARCHIVED
          nullable: true
        buyable:
          type: boolean
          nullable: true
        buyerCanCancel:
          type: boolean
          nullable: true
        perks:
          type: array
          nullable: true
          items:
            type: string
        linkedPlatforms:
          type: array
          description: External platforms where this product is linked/published.
          items:
            $ref: '#/components/schemas/LinkedPlatformSync'
...
    BillingInterval:
      type: string
      description: Unit used for billing, duration and free trial.
      enum:
      - DAY
      - WEEK
      - MONTH
      - YEAR
    PricingModel:
      type: string
      description: Pricing model, aligned with Wix Plan.pricingVariants[].pricingModel.
      enum:
      - SUBSCRIPTION
      - SINGLE_PAYMENT_FOR_DURATION
      - SINGLE_PAYMENT_UNLIMITED
    PricingDuration:
      type: object
      description: Generic duration structure used for billingCycle, duration and
        freeTrial.
      required:
      - interval
      - count
      properties:
        interval:
          $ref: '#/components/schemas/BillingInterval'
        count:
          type: integer
          minimum: 1
    PricingVariantConfig:
      type: object
      description: Pricing configuration for the product, conceptually matching Wix
        Plan.pricingVariants[0].
      required:
      - pricingModel
      - flatRate
      properties:
        pricingModel:
          $ref: '#/components/schemas/PricingModel'
        flatRate:
          type: number
          format: decimal
        billingCycle:
          $ref: '#/components/schemas/PricingDuration'
          nullable: true
        duration:
          $ref: '#/components/schemas/PricingDuration'
          nullable: true
        freeTrial:
          $ref: '#/components/schemas/PricingDuration'
          nullable: true
    LinkedPlatformSync:
      type: object
      description: Synchronization state of this product on an external platform.
      required:
      - platformName
      - isSynced
      properties:
        platformName:
          type: string
          description: Name of the external platform (e.g. "wix", "shopify", "webflow").
          example: wix
        idOnPlatform:
          type: string
          nullable: true
          description: Identifier of the product/plan on the external platform. Null
            if it has not been created there yet.
        revision:
          type: string
          nullable: true
          description: Revision or version identifier as used by the platform (e.g.
            Wix plan.revision).
        visibilityOnPlatform:
          type: string
          nullable: true
          description: Visibility/status of the product on that platform.
          enum:
          - DRAFT
          - PUBLISHED
          - ARCHIVED
          - HIDDEN
          - NOT_PUBLISHED
        isSynced:
          type: boolean
          description: True if the last local change has been successfully pushed
            to this platform.
        isSourceOfTruth:
          type: boolean
          default: false
          description: If true, this platform is considered the authoritative source
            for product data; sync direction is primarily platform -> local.
        lastSyncedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the last successful sync to this platform.
        syncError:
          type: string
          nullable: true
          description: Last error message when attempting to sync with this platform,
            if any.
        hasLocalChanges:
          type: boolean
          default: false
          description: True if there are local changes that haven't been pushed to the platform yet.
        hasIncomingChanges:
          type: boolean
          default: false
          description: True if there are changes on the platform that haven't been pulled locally yet.
        localHash:
          type: string
          nullable: true
          description: Hash of the local product state for change detection.
        remoteHash:
          type: string
          nullable: true
          description: Hash of the remote product state for change detection.